requirements.txt
Django
psycopg2-binary
pybase62